Added service to get beneficiary info

+++ src/main/java/com/banking/controller/beneficiary/BeneficiaryController.java
@@ -33,10 +33,10 @@ public class BeneficiaryController {
 
     @RequestMapping(value = "/getBeneficiary/{clientId}", method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
-    public ResponseEntity<BeneficiaryReponse> getBeneficiary(@PathVariable(value = "clientId") long clientId) {
-        logger.info("Getting Request to get beneficiary info for client : {} ", clientId);
-        BeneficiaryReponse beneficiaryReponse  = beneficiaryProcessor.getBeneficiaryInfo(clientId);
-        return new ResponseEntity<>(beneficiaryReponse, HttpStatus.OK);
+    public ResponseEntity<BeneficiaryReponse> getBeneficiary(@PathVariable(value = "clientId") int clientId) {
+            logger.info("Getting Request to get beneficiary info for client : {} ", clientId);
+            BeneficiaryReponse beneficiaryReponse  = beneficiaryProcessor.getBeneficiaryInfo(clientId);
+            return new ResponseEntity<>(beneficiaryReponse, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/updateBeneficiary", method = RequestMethod.POST, produces = "application/json")
